Barcode Implementation Fixes #1
Initial creation of necessary classes and imports for barcode
scanner/reader and use of camera. Initial format is based off of the API
sample; Android Vision. Still working out bugs to get to a runnable test
state. This initial shell will be modified over time to integrate into
the app as needed.